Add input and canvas size feature in css and js
This commit adds a CSS variable for sizing the pixel class within the
div container. There was a good bit of JS commited as well for handling
user input for the grid size, creating a new canvas sized to the user
input, and passing the user input to the css variable to handle the new
size for the pixel class.